[video] Video improvements omnibus
- Add a new decoder implementation that streams data on the fly with libav
- Generate soundfile processes when dropping a video file
- Many many small fixes to timing, seeking, memory leaks, etc
- Less memory used, projects with video will load faster

+++ src/plugins/score-plugin-media/Media/RMSData.cpp
@@ -179,7 +179,7 @@ RMSData::frame(int64_t start_frame, int64_t end_frame) const noexcept
   {
     for(std::size_t k = 0; k < channels; k++)
     {
-      const int64_t val = begin[k];
+      const float val = begin[k];
       sum[k] = abs_max(sum[k], val);
     }
   }
